requirements.txt
absl-py==1.3.0
asgiref==3.5.2
attrs==22.1.0
click==8.1.3
cmake==3.22.4
contourpy==1.0.6
cycler==0.11.0
Django==4.1.3
djangorestframework==3.14.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
mediapipe==0.8.10
numpy==1.21.6
opencv-contrib-python-headless==4.6.0.66
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==9.3.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
gunicorn